requirements.txt
Flask
PyPDF2
reportlab
pdfminer.six